moderation-agent/requirements.txt
functions-framework==3.*
google-generativeai>=0.8.0
flask
